fix: use correct `//# sourceURL` with `find_additional_modules`
Previously, the automatically inserted `//# sourceURL` comments for
modules found using `find_additional_modules` referenced the
temporary directory, not the original source directory the modules
were from. This change makes sure they use the correct on-disk path,
ensuring source maps are resolved correctly, and breaking debugging
in IDEs works.
