+++ packages/wrangler/src/dev/remote.tsx
@@ -1,6 +1,4 @@
-import { readFile } from "node:fs/promises";
 import path from "node:path";
-import { pathToFileURL } from "url";
 import { Text } from "ink";
 import SelectInput from "ink-select-input";
 import React, { useState, useEffect, useRef } from "react";
@@ -8,6 +6,7 @@ import { useErrorHandler } from "react-error-boundary";
 import { helpIfErrorIsSizeOrScriptStartup } from "../deploy/deploy";
 import { printBundleSize } from "../deployment-bundle/bundle-reporter";
 import { getBundleType } from "../deployment-bundle/bundle-type";
+import { withSourceURLs } from "../deployment-bundle/source-url";
 import { logger } from "../logger";
 import { syncAssets } from "../sites";
 import {
@@ -506,17 +505,6 @@ export async function getRemotePreviewToken(props: RemoteProps) {
 	});
 }
 
-/**
- * Adds `//# sourceURL` comment so V8 knows where source files are on disk.
- * This URL is returned in `Debugger.scriptParsed` events, ensuring inspector
- * clients resolve source mapping URLs correctly. It also appears in stack
- * traces, allowing users to click through to where errors are thrown. Note,
- * Miniflare includes similar code, so we only need to do this in remote mode.
- */
-function withSourceURL(source: string, sourcePath: string) {
-	return `${source}\n//# sourceURL=${pathToFileURL(sourcePath)}`;
-}
-
 async function createRemoteWorkerInit(props: {
 	bundle: EsbuildBundle;
 	modules: CfModule[];
@@ -532,9 +520,9 @@ async function createRemoteWorkerInit(props: {
 	compatibilityFlags: string[] | undefined;
 	usageModel: "bundled" | "unbound" | undefined;
 }) {
-	const content = withSourceURL(
-		await readFile(props.bundle.path, "utf-8"),
-		props.bundle.path
+	const { entrypointSource: content, modules } = withSourceURLs(
+		props.bundle.path,
+		props.modules
 	);
 
 	// TODO: For Dev we could show the reporter message in the interactive box.
@@ -556,21 +544,6 @@ async function createRemoteWorkerInit(props: {
 		undefined
 	); // TODO: cancellable?
 
-	const modules = props.modules.map((module) => {
-		// If this is a JavaScript module with a `filePath`, add a `//# sourceURL`
-		// comment so source maps resolve correctly
-		if (
-			module.filePath !== undefined &&
-			(module.type === "esm" || module.type === "commonjs")
-		) {
-			// `module.content` may be a `Buffer`
-			let newContent = module.content.toString();
-			newContent = withSourceURL(newContent, module.filePath);
-			return { ...module, content: newContent };
-		} else {
-			return module;
-		}
-	});
 	if (assets.manifest) {
 		modules.push({
 			name: "__STATIC_CONTENT_MANIFEST",
